# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "x402 Agent Toolkit",
    "version": "1.0.0",
    "description": "Paid utility endpoints for AI agents. Pay with USDC on Base or Solana. Pure computation, no LLM dependency.",
    "contact": { "email": "sna4an@users.noreply.github.com" },
    "x-guidance": "This API provides data processing utilities for AI agents. All endpoints require USDC payment via the x402 protocol. Use POST with JSON body for all endpoints."
  },
  "servers": [{ "url": "https://x402-agent-toolkit.vercel.app" }],
  "paths": {
    "/api/url-to-text": {
      "post": {
        "operationId": "urlToText",
        "summary": "Scrape URL and extract text",
        "description": "Fetches a URL and returns clean text content with HTML stripped. Supports articles, docs, and web pages.",
        "x-payment-info": {
          "price": { "mode": "fixed", "currency": "USD", "amount": "0.020000" },
          "protocols": [{ "x402": {} }]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": { "type": "string", "minLength": 1, "description": "URL to scrape" }
                },
                "required": ["url"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scraped text content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": { "type": "string" },
                    "text": { "type": "string" },
                    "chars": { "type": "integer" },
                    "fetchedAt": { "type": "string" }
                  }
                }
              }
            }
          },
          "402": { "description": "Payment Required" }
        }
      }
    },
    "/api/json-transform": {
      "post": {
        "operationId": "jsonTransform",
        "summary": "Transform JSON data",
        "description": "Apply transformations to JSON: filter by key/value, pick specific keys, flatten nested objects, list keys, or count items.",
        "x-payment-info": {
          "price": { "mode": "fixed", "currency": "USD", "amount": "0.010000" },
          "protocols": [{ "x402": {} }]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": { "description": "Input JSON data (object or array)" },
                  "operation": { "type": "string", "enum": ["filter", "pick", "flatten", "keys", "count"], "description": "Transform operation" },
                  "params": { "type": "object", "description": "Operation parameters (e.g. {key, value} for filter, {keys:[]} for pick)" }
                },
                "required": ["data", "operation"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transformed result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "operation": { "type": "string" },
                    "result": { "description": "Transformed data" }
                  }
                }
              }
            }
          },
          "402": { "description": "Payment Required" }
        }
      }
    },
    "/api/csv-parse": {
      "post": {
        "operationId": "csvParse",
        "summary": "Parse CSV to JSON",
        "description": "Parse CSV text into structured JSON with auto-detected delimiters and headers. Supports quoted fields.",
        "x-payment-info": {
          "price": { "mode": "fixed", "currency": "USD", "amount": "0.010000" },
          "protocols": [{ "x402": {} }]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "csv": { "type": "string", "minLength": 1, "description": "CSV text to parse" },
                  "delimiter": { "type": "string", "description": "Optional delimiter (auto-detected if omitted)" },
                  "hasHeader": { "type": "boolean", "default": true, "description": "Whether first row is headers" }
                },
                "required": ["csv"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parsed CSV data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "delimiter": { "type": "string" },
                    "headers": { "type": "array", "items": { "type": "string" } },
                    "rows": { "type": "integer" },
                    "data": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          },
          "402": { "description": "Payment Required" }
        }
      }
    },
    "/api/text-stats": {
      "post": {
        "operationId": "textStats",
        "summary": "Get text statistics",
        "description": "Analyze text and return statistics: word count, character count, sentence count, reading time, top words, and more.",
        "x-payment-info": {
          "price": { "mode": "fixed", "currency": "USD", "amount": "0.005000" },
          "protocols": [{ "x402": {} }]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": { "type": "string", "minLength": 1, "description": "Text to analyze" }
                },
                "required": ["text"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Text statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "characters": { "type": "integer" },
                    "words": { "type": "integer" },
                    "sentences": { "type": "integer" },
                    "paragraphs": { "type": "integer" },
                    "readingTimeMinutes": { "type": "number" },
                    "topWords": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          },
          "402": { "description": "Payment Required" }
        }
      }
    },
    "/api/hash-generate": {
      "post": {
        "operationId": "hashGenerate",
        "summary": "Generate cryptographic hash",
        "description": "Generate SHA256, MD5, SHA512, SHA1, or SHA384 hash from any input string.",
        "x-payment-info": {
          "price": { "mode": "fixed", "currency": "USD", "amount": "0.005000" },
          "protocols": [{ "x402": {} }]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "input": { "type": "string", "minLength": 1, "description": "String to hash" },
                  "algorithm": { "type": "string", "enum": ["sha256", "md5", "sha512", "sha1", "sha384"], "default": "sha256", "description": "Hash algorithm" }
                },
                "required": ["input"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated hash",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "algorithm": { "type": "string" },
                    "hash": { "type": "string" },
                    "inputLength": { "type": "integer" }
                  }
                }
              }
            }
          },
          "402": { "description": "Payment Required" }
        }
      }
    }
  }
}